pyproject.toml
[project]
name = "engfleet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.8",
    "google-adk[eval]>=1.15.1",
    "poml>=0.0.8",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ruff>=0.13.3",
]

[tool.ruff]
exclude = [
    "ruff/",
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "ref"
]

[tool.ruff.lint]
select = [
    # Pyflakes
    "F",
    # pycodestyle
    "E",
    # isort
    "I",
    # pep8-naming
    "N",
    # pyupgrade
    "UP",
    # flake8-2020
    "YTT",
    # flake8-async
    "ASYNC",
    # flake8-bandit
    "S506",  # unsafe-yaml-load
    # flake8-bugbear
    "B",
    # flake8-executable
    "EXE",
    # flake8-pie
    "PIE",
    # flake8-pyi
    "PYI",
    # flake8-simplify
    "SIM",
    # pylint
    "PLE",      # errors
    "PLW",      # warnings
    "PLR1711",  # useless-return
    # Ruff-specific rules
    "RUF",
]
ignore = [
    # pydocstyle
    "D105",  # undocumented-magic-method
    "D107",  # undocumented-public-init
    "D203",  # one-blank-line-before-class
    "D205",  # blank-line-after-summary
    "D212",  # multi-line-summary-first-line
    "D401",  # non-imperative-mood
    # Pyflakes
    "F403",  # undefined-local-with-import-star
    "F841",  # unused-variable
    # pycodestyle
    "E402",  # module-import-not-at-top-of-file
    "E501",  # line-too-long
    # pep8-naming
    "N802",  # invalid-function-name
    # pylint warnings
    "PLW3201",  # bad-dunder-method-name
    # flake8-pyi
    "PYI041",  # redundant-numeric-union
    # pyupgrade
    "UP015",  # redundant-open-modes
    "UP032",  # f-string
    # flake8-bugbear
    "B011",  # assert-false
    # Ruff-specific rules
    "RUF028",  # invalid-formatter-suppression-comment
    # flake8-simplify
    "SIM102",  # collapsible-if
    "SIM108",  # if-else-block-instead-of-if-exp
    "SIM117",  # multiple-with-statements
]
task-tags = ["TODO", "FIXME", "XXX", "@@@"]

[tool.ruff.lint.isort]
lines-after-imports = 2
section-order = [
    "future",
    "standard-library",
    "third-party",
    "internal",
    "first-party",
    "local-folder",
]
